[lldb] Explicitly cast bit-field to uint64_t when creating optional
llvm::Optional is now defined as `std::optional`. While it would be
great to remove llvm::Optional and use std::optional everywhere, that is
a non-goal of this patch.

The std::optional constructor cannot accept non-const bit-fields. To
work around this, we explicitly create a new uint64_t out of the
bit-field and then use that to create the optional.

(cherry picked from commit 5a86013d8172928de2556c7061132403c799861e)

+++ lldb/source/Plugins/TypeSystem/Swift/TypeSystemSwiftTypeRef.cpp
@@ -2579,7 +2579,7 @@ TypeSystemSwiftTypeRef::GetBitSize(opaque_compiler_type_t type,
         /// Type::GetByteSize() may call into this function again.
         llvm::Optional<uint64_t> GetStaticByteSize() {
           if (m_byte_size_has_value)
-            return m_byte_size;
+            return uint64_t(m_byte_size);
           return {};
         }
       };
